search_demo/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
torch>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4

